add protected route
