+++ packages/client/index.html
@@ -8,7 +8,7 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/App.tsx"></script>
+    <script type="module" src="/src/index.tsx"></script>
     <script src="./sw.js"></script>
   </body>
 </html>
