--- packages/client/src/pages/index.tsx
@@ -1,18 +1,0 @@
-import { useEffect } from 'react'
-import { Navigate } from 'react-router-dom'
-
-function App() {
-  useEffect(() => {
-    const fetchServerData = async () => {
-      const url = `http://localhost:${__SERVER_PORT__}`
-      const response = await fetch(url)
-      const data = await response.json()
-      console.log(data)
-    }
-
-    fetchServerData()
-  }, [])
-  return <Navigate to={'/game'} replace />
-}
-
-export default App
